Upload term_project
This is the "zeze's restaurant" menu ordering program I created with Python's tkinter library. When a user puts his/her brain out and places an order, it is designed to guide the user in the right direction.

The interesting thing about the program I created is that birthday discounts are available. When the orderer's birthday is entered, a 10% birthday discount is applied when it matches the computer time.